FM-1017 Use updated CAS terminology across SAR code- #5117
This commit updates SAR code to use updated and consistent SAR terminology

There is no change to the template output

In a subsequent commit we’ll use the CAS terminology in the template itself, but as this changes the template (which triggers SAR change control) that will be done on a branch with additional template changes
