+++ src/test/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/cas2hdc/integration/sar/Cas2HdcSarComplianceTest.kt
@@ -1,44 +1,43 @@
-package uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2v2.integration.sar
+package uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2hdc.integration.sar
 
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model.ServiceName
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2.model.Cas2ServiceOrigin
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2hdc.integration.sar.Cas2SarTestBase
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2hdc.jpa.entity.Cas2Cohort
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.integration.givens.givenAnOffender
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.integration.sar.CasSarFixtureAsserter
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.DomainEventType
 import java.time.LocalDate
 
 /**
- * Per-service SAR compliance test for CAS2V2.
+ * Per-service SAR compliance test for CAS2.
  *
  * Cross-service SAR infrastructure (Flyway schema, JPA entity snapshot, template
  * endpoint smoke tests) lives in `SarIntegrationTest` — they only need to run
  * once for the whole app, since all four CAS services share one DB and one
  * template file.
  *
- * This class verifies CAS2V2's slice end-to-end against CAS2V2-specific fixtures
+ * This class verifies CAS2's slice end-to-end against CAS2-specific fixtures
  * via [CasSarFixtureAsserter].
  */
-class Cas2v2SarComplianceTest : Cas2SarTestBase() {
+class Cas2HdcSarComplianceTest : Cas2HdcSarTestBase() {
 
   companion object {
-    const val TEST_CRN = "X320743"
-    const val TEST_NOMS_NUMBER = "A1234BE"
+    const val TEST_CRN = "X320742"
+    const val TEST_NOMS_NUMBER = "A1234BD"
     const val TEST_NOMIS_USER_NAME = "SAR-TEST-NOMIS-USER"
+    const val TEST_EXTERNAL_USER_NAME = "SAR-TEST-EXTERNAL-USER"
     const val TEST_ASSESSOR_NAME = "SAR-TEST-ASSESSOR"
-    const val TEST_NACRO_REFERRAL_ID = "0000000002"
+    const val TEST_NACRO_REFERRAL_ID = "0000000001"
     const val TEST_REFERRING_PRISON_CODE = "ABC"
     const val TEST_TELEPHONE_NUMBER = "0123456"
     val TEST_FROM_DATE: LocalDate = LocalDate.of(2019, 1, 1)
     val TEST_TO_DATE: LocalDate = LocalDate.of(2024, 12, 31)
 
-    const val EXPECTED_API_RESPONSE_PATH = "/sar/cas2v2-expected-api-response.json"
-    const val EXPECTED_REPORT_PATH = "/sar/cas2v2-expected-report.html"
-    const val GENERATED_API_RESPONSE_FILENAME = "sar/cas2v2-expected-api-response.json"
-    const val GENERATED_REPORT_FILENAME = "sar/cas2v2-expected-report.html"
+    const val EXPECTED_API_RESPONSE_PATH = "/sar/cas2-hdc-expected-api-response.json"
+    const val EXPECTED_REPORT_PATH = "/sar/cas2-hdc-expected-report.html"
+    const val GENERATED_API_RESPONSE_FILENAME = "cas2-hdc-sar-api-response.json.log"
+    const val GENERATED_REPORT_FILENAME = "cas2-hdc-sar-report.html.log"
   }
 
   private val asserter by lazy {
@@ -68,41 +67,41 @@ class Cas2v2SarComplianceTest : Cas2SarTestBase() {
         withNomsNumber(TEST_NOMS_NUMBER)
       },
     )
-    val user = cas2NomisUserEntity(Cas2ServiceOrigin.BAIL, name = TEST_NOMIS_USER_NAME)
+    val user = cas2NomisUserEntity(Cas2ServiceOrigin.HDC, name = TEST_NOMIS_USER_NAME)
+    val assessor = cas2ExternalUserEntity(name = TEST_EXTERNAL_USER_NAME)
     val application = cas2ApplicationEntity(
       offenderDetails,
       user,
-      Cas2ServiceOrigin.BAIL,
+      Cas2ServiceOrigin.HDC,
       referringPrisonCode = TEST_REFERRING_PRISON_CODE,
       telephoneNumber = TEST_TELEPHONE_NUMBER,
-      data = CAS2V2_APPLICATION_DATA,
+      data = CAS2_APPLICATION_DATA,
       document = "null",
     )
     cas2ApplicationEntity(
       offenderDetails,
       user,
-      Cas2ServiceOrigin.BAIL,
+      Cas2ServiceOrigin.HDC,
       referringPrisonCode = TEST_REFERRING_PRISON_CODE,
       telephoneNumber = TEST_TELEPHONE_NUMBER,
-      data = CAS2V2_APPLICATION_DATA,
-      document = CAS2V2_APPLICATION_DOCUMENT,
-      cohort = Cas2Cohort.ATCR,
+      data = CAS2_APPLICATION_DATA,
+      document = CAS2_APPLICATION_DOCUMENT,
     )
     val assessment = cas2AssessmentEntity(
       application,
-      Cas2ServiceOrigin.BAIL,
+      Cas2ServiceOrigin.HDC,
       assessorName = TEST_ASSESSOR_NAME,
       nacroReferralId = TEST_NACRO_REFERRAL_ID,
     )
 
     cas2ApplicationNoteEntity(application, assessment, user)
-    val statusUpdate = cas2StatusUpdateEntity(application, assessment, user)
+    val statusUpdate = cas2StatusUpdateEntity(application, assessment, assessor)
     cas2StatusUpdateDetailEntity(statusUpdate)
-    domainEventEntity(offenderDetails, application.id, assessment.id, null, DomainEventType.CAS2_APPLICATION_SUBMITTED, ServiceName.cas2v2)
+    domainEventEntity(offenderDetails, application.id, assessment.id, null, DomainEventType.CAS2_APPLICATION_SUBMITTED, ServiceName.cas2)
   }
 
   @Test
-  fun `CAS2V2 SAR API should return expected data`() {
+  fun `CAS2 HDC SAR API should return expected data`() {
     setupTestData()
     asserter.assertApiDataMatchesFixture(
       crn = TEST_CRN,
@@ -112,7 +111,7 @@ class Cas2v2SarComplianceTest : Cas2SarTestBase() {
   }
 
   @Test
-  fun `CAS2V2 SAR report should render as expected`() {
+  fun `CAS2 HDC SAR report should render as expected`() {
     setupTestData()
     asserter.assertReportMatchesFixture(
       crn = TEST_CRN,
